Implemented input polling.

+++ Matter/src/Matter/ImGui/ImGuiLayer.cpp
@@ -188,7 +188,7 @@ namespace Matter {
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e) {
 
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2((float)e.GetWidth(), (float)e.GetHeight());
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		glViewport(0, 0, e.GetWidth(), e.GetHeight());
